Add files via upload
Big Data and SQL do not have a lot in common. However, over the last couple of years this changed, and more and more people want to integrate the data from their Big Data systems into their SQL data warehouses. The most important technologies in the Big Data space are Spark as technology itself and Databricks as a PaaS solution hosting it. These new tools may be frightening in the beginning but once you get to know them you will realize that they are quite similar to your regular SQL tools. And this is what this session is about - giving a regular SQL developer insight into Big Data and showing how SQL can still be used to do Big Data processing with Spark and Databricks